AUW Fix, Held Items, Two commands
AUW compat was broken, I'm glad I didn't roll this update out.
Added Held Item temperature, for stuff in your mainhand or offhand. Can be configured and has a command.
Added two other commands, one for help on commands and the other to reload a mod's built-in compatibility JSON (good to use if you add a compatible mod at a later date)