0.2.4 Crash Fix for Servers (oops)
I guess this mod never worked on servers for the longest time. That's what I get for assuming annotations work like I expect.

Come to think of it that's probably what happened, the annotation removed the field, but the assignment to the variable did not get removed as it'd get transferred over to the static {} area

god