Swap to Identity JSON for NBT, JSON Migration, beginning of Rustic, some cleanup
New JSON handling to get started with NBT options. Added automatic migration to move from the old format to the new.
This is to allow for NBT matching, which is currently untested but will be used for Rustic compatibility and other stuff.

Should have set this up sooner, but JSON is a constant learning process!