Baubles, Potion Duration, ArmorTemperatures Rework
Removed that blocks config area, it's all in miscellaneous now.

Potion duration can be configured.

Baubles compatibility added. Bauble information is stored in armorTemperatures.json, and /sd addArmor works with any item now.

The format of armorTemperatures.json was changed to use the JsonItemIdentity system, so the mod will try to migrate the old format to the new.

I can't wait to be able to remove all this migration stuff...